refactor: extract notes formatting
Jira: AI-516
